Spree Navigation Component & Structure Migration, closes #33
migrate to component structure

remove dev line
