+++ packages/frontend/addon/components/spree-address.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import layout from '../templates/components/spree-address';
+
+export default Ember.Component.extend({
+  layout: layout
+});
