+++ packages/frontend/addon/components/spree-details.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import layout from '../templates/components/spree-details';
+
+export default Ember.Component.extend({
+  layout: layout
+});
