+++ packages/frontend/addon/components/spree-lineitems.js
@@ -1,0 +1,14 @@
+import Ember from 'ember';
+import layout from '../templates/components/spree-lineitems';
+
+export default Ember.Component.extend({
+  layout: layout,
+  action: 'deleteLineItem',
+
+  actions: {
+    deleteLineItem: function(lineItem) {
+      this.sendAction('action', lineItem);
+      return false;
+    }
+  }
+});
