+++ packages/frontend/addon/components/spree-navigation.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import layout from '../templates/components/spree-navigation';
+
+export default Ember.Component.extend({
+  layout: layout
+});
