+++ packages/frontend/addon/components/spree-properties.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import layout from '../templates/components/spree-properties';
+
+export default Ember.Component.extend({
+  layout: layout
+});
