+++ packages/frontend/addon/components/spree-shipment.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import layout from '../templates/components/spree-shipment';
+
+export default Ember.Component.extend({
+  layout: layout
+});
