+++ packages/frontend/addon/components/spree-source.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import layout from '../templates/components/spree-source';
+
+export default Ember.Component.extend({
+  layout: layout
+});
