+++ packages/frontend/addon/templates/components/spree-details.hbs
