+++ packages/frontend/addon/templates/components/spree-images.hbs
