+++ packages/frontend/addon/templates/components/spree-lineitems.hbs
