+++ packages/frontend/addon/templates/components/spree-navigation.hbs
@@ -1,0 +1,23 @@
+<nav class="top-bar">
+  <section class="top-bar-section">
+    <ul>
+      <li>
+        {{#link-to 'spree.index'}}
+          Spree Ember
+        {{/link-to}}
+      </li>
+    </ul>
+    <ul class="right">
+      <li>
+        {{#link-to 'spree.cart'}}
+          Cart: &nbsp;
+          {{#if spree.currentOrder}}
+            {{spree.currentOrder.itemCount}}
+          {{else}}
+             0
+          {{/if}}
+        {{/link-to}}
+      </li>
+    </ul>
+  </section>
+</nav>
