+++ packages/frontend/addon/templates/components/spree-properties.hbs
