+++ packages/frontend/addon/templates/components/spree-source.hbs
