+++ packages/frontend/app/components/spree-images.js
@@ -1,21 +1,3 @@
-import Ember from 'ember';
+import spreeImages from 'ember-cli-spree-frontend/components/spree-images';
 
-export default Ember.Component.extend({
-  activeImage: null,
-
-  selectedImage: Ember.computed('activeImage', 'images.@each', function() {
-    var activeImage = this.get('activeImage');
-    if (activeImage) {
-      return activeImage;
-    } else {
-      return this.get('images.firstObject');
-    }
-  }),
-
-  actions: {
-    selectImage: function(image) {
-      this.set('activeImage', image);
-      return false;
-    }
-  }
-});
+export default spreeImages;
