+++ packages/frontend/app/components/spree-navigation.js
@@ -1,0 +1,3 @@
+import spreeNavigation from 'ember-cli-spree-frontend/components/spree-navigation';
+
+export default spreeNavigation;
