+++ packages/frontend/app/templates/spree.hbs
@@ -1,0 +1,2 @@
+{{spree-navigation}}
+{{outlet}}
