+++ packages/frontend/app/templates/spree/index.hbs
@@ -1,6 +1,8 @@
 <div class="row">
   <div class="small-12 columns text-center">
     <h1>Welcome to Spree Ember</h1>
+    <h6><strong>Note:</strong> You can delete the "Welcome to Ember.js" message in
+      your application.hbs file.</h6>
     {{#link-to 'spree.products.index'}}
       <button>Products</button>
     {{/link-to}}
